+++ clang-tidy/misc/MiscTidyModule.cpp
@@ -27,6 +27,7 @@
 #include "UnusedAliasDeclsCheck.h"
 #include "UnusedParametersCheck.h"
 #include "UnusedUsingDeclsCheck.h"
+#include "VirtualFunctionDeclarationCheck.h"
 
 namespace clang {
 namespace tidy {
@@ -68,6 +69,8 @@ public:
         "misc-unused-parameters");
     CheckFactories.registerCheck<UnusedUsingDeclsCheck>(
         "misc-unused-using-decls");
+    CheckFactories.registerCheck<VirtualFunctionDeclarationCheck>(
+        "misc-virtual-function-declaration");
   }
 };